feat(Permissions): Added deletepermission command
Added deletepermission command to delete
permissions by their ids.
This is needed to delete permissions which
genuine delete commands (like removerolemanager)
are too complex, or if the permission is
corrupted so cannot be
accessible by an object query.
Also removed insertDiscordNext function,
because it duplicates ids if called within a cycle.
Perhaps adding some locks inside the DB manager
is needed to use such functions in the future.
For now - just let the particular commands
to increment the ids on their own.