fix(5): Fixed scanning empty time defintions
The Bot considered an empty string as a valid
year definition, because there was no check
against an empty string in the definition parser.
So in such cases the year value was set as NaN
and the reminder got immediatelly deleted.
Fixed by adding the non-empty check in the TimeArg class.