feat(images): Add advanced limitations for images
Added limitations on the complexity and length of image templates.
Added limitations on the dimensions of the base image
in templates.
Made the max font file size configurable for the Bot.
Made the target JPG quality configurable (95 by default).
In the makeimage command moved validation out
of the executeForDiscord function to keep the architecture intact.
Moved functions related to downloading of files
from ImageGenerator to BotUtils.
Added function to calculate JS objects complexity to BotUtils.
Minor code style fixes.