fix(#24): Make fetching guild structure optional
Added new copy_org_structure_locally preference.
If it's not set then the Bot will not store channels, members
and roles locally on each command and on start up..
Also removed the dependency on the local members, channels
and roles DB tables when checking command arguments.
This will allow to work without copying the guilds structure.
The mechanism of copying guild structures will remain
in the source code in case it is needed in the future.