feat(reminders): made proper response on adding
Changed the response to the "remind" command to include
the reminder to be scheduled, and the result of scheduling
(e.g. if a reminder was scheduled for the past then
it will be missing from the second part of the output).
Reminders are printed with all the details.
Changed the text strings accordingly.
Added async/await keywords to some operations with tasks
to be able to get the result of scheduling immediatelly
in the same command.
Changed the e2e tests related to reminders accordingly.