fix(#25): Make safe charset for censoring
Replaced the '$' sign by the '?' sign in the char template
for euphemisms. '$' caused problems with regexes
because it inserts the string to be replaced back
into the result string if it occurs together with
other special symbols in certain combinations
(that's accordinf to the specs of string.replace function).
For example, if the euphemism contains '$&' combination
then the bad word will be put back to the string
instead of the '$&' combination.