feat(images): Implement #32, dynamic template parameters
Added possibility to define the set of customizable style
attributes for each template.
Each template can have a "style" sub-object with properties,
which have names and default values.
For each such property it's possible to set a value while
writing the makeimage command for the Bot inside
a new "style" parameter.
In this case the Bot will take the dynamic value from
the command instead of a hardcoded value.
If the dynamic value is not specified, then the default value will be used.
Almost all layer attributes can be customized in such way.
Also, the base width, height and xShift, yShift can be customized.
To be able to pass the complex "style" parameter,
a new scanner for arbitrary objects was added.
Also, the old 3 parameters for the makeimage command
were removed (xDiif, yDiff, fontSize), as they can be dynamically set
the same way as other attributes.
Also, loosen the template validation a bit to be able to
use strings as the base width and height.
Finally, commented out 2 links in the web-api, as it blocks the launch.
(to be investigated later).