feat(web), fix(#38): basic web commands execution
Implemented execution of the most simple Bot commands from Web UI.
Replaced pure divs by forms in the arguments UI.
Added backend support for commands parsing and execution.
The backend part is implemented as similar as possible to
the original Discord part.
Added the multer lib to handle form data from the frontend accordignly.
Rewritten some stuff related to permissions in order not to rely
on the native Discord messages (which are not available from Web UI).
Web interface now uses the same locale settings as set
for the Bot in Discord client.
The Bot messages now store the reply content inside the object
if it's not possible to send it to the original source. This is used
to forward the reply to the frontend.
Fixed a bug in settings and mysettings command which prevented
users from asking for particular settings.  The root causes were:
1. Wrong scanner used (not the full string one)
2. Wrong function used to check if the setting exists.
Note that the most commands still do not work from Web UI.