feat(android) feat(suggestions): Add support of input suggestions
- Added support of suggestions for arguments input in UI.
- The suggestions are retrieved from the server using
special "suggestions" commands when the UI app
(e.g. Android) build the views for the command arguments.
- The suggestions commands extend the base
DiscordCommand class, and some of them may extend
particular commands to reuse the functionality whenever
possible.
- Added ability for commands to put suggestions to
the output before responding to the Web client.
- For the regular commands which suggestion commands
are based on, integrated the procedure of making
suggestion into the main execution function
to reuse the functionality.
- Added UI type of scanners (for Web and Android), so
the clients can build UI accordingly.
- Implemented scanners in Android for simple string
and array arguments. The rest of scanner types just
use a simple EditText (to be implemented later).
- The array scanner uses a chip group to represent already
entered values.
- Added support of suggestions to the simple and array
scanners (in a form of spinner with an auto-complete
text view).
- Changed the UI theme from AppCompat to Material to
use some advanced views in the UI.
- Changed the input type of "settings" and "mysettings"
commands to array, so several settings can be
retreived at once.
- Refactored Bot's command handlers to reuse the
functionality for different sources in the future,
and also to reuse the functions for both regular
and suggestions commands.
- Added the list of possible suggestion commands
to the Discord command handler.
- Added suggestions controller and endpoint
to the Web API.
- Modified the scanning procedure on the server to
consider the scanning type (sequential or by name).
In general, while scanning arguments "by name" the
scanner will behave more greedy. For example,
the simple string scanner will take the whole provided
text instead of a substring until the next non-quoted
space.
This is done because UI clients also provide arguments
by their names, not sequentially, so the greediness on one
argument will not do any harm to the other arguments.
- Added possiblity to declare corresponding suggestion
commands either in particular arguments or in scanner
classes. For example, channel suggestions will be helpful
for all channel arguments, hence we declare it in the
channel array scanner type instead of declaring for each
particular channel argument in corresponsing commands.
- Declared appropriate suggestions commands for
existing arguments and scanner types.
- The suggestions on mentions (users, channels, roles)
are made based on the guild info, so are subjects
to changes and limitations in the future.
- For the mentions scanners (users, channels and roles)
added possibility to scan user-friendly mentions
(e.d. @David instead of <@57354873463587>).
This is needed because the UI clients (Web and Android)
provide user-friendly names instead of ids for
the argument inputs.
- Other minor improvements.