feat(Android) feat(suggestions): Add org-wide suggestions
Made a list of suggestion commands, which can be used to
retrieve org-wide suggestions. Those can be re-used across
multiple commands and arguments (if only their assigned
suggestions command id matches to the org-wide cache).
Extended the Org table in the DB to keep the list of
suggestion commands (subject to change - see TODO notes).
Fetching the org-wide suggestions starts on launching the
Modules fragment (that is, as soon as an org is selected).
If the app does not find suggestions in the org-wide cache,
then it gets it using the normal procedure for a particular
command argument.
Also, set a suggestion command for the ids argument
of the delete-permission-command.