feat(Telegram) feat(sources): Add basic support for Telegram
Added the very basic support for Telegram platform.
The "Telegraf" library is used for the implementation.
Only "ping" command is supported for the Telegram Bot for now.
Renamed "commands_discord" folder to "command", because
it was decided to support distinguishing of sources inside
the command classes. Hence each command class will be
a universal handler for all supported sources (Discord, Telegram),
if applicable, of course.
Made command-handling functions source-independent
(in some managers and handler classes), including the Web-parser.
Locale, DB and permission management is still not
implemented for Telegram.
Fixed a typo in the bad words E2E tests.