feat(Telegram): Extend the commands support
IMPORTANT: Updated NodeJS to 16.13.1 in order to
support the latest Telegraf version.
Added permission system for Telegram, similar to
what is implemented for Discord. Particular permissions
are not implemented yet though, the Bot only
distinguishes admin and non-admin users for now.
User in a private channel is always considered as admin.
Added support of LangManager for Telegram.
Added support of various commands for
Telegram: image making commands, some
settings commands, the help command.
For the makeimage command: added a delay of 3
seconds after making the picture and before sending
it as the await does not seem to work well in this case.
Made some message/command functions
source-abstract, so they can be called safely even
without knowing what is the source of the message:
e.g. reading text attachments, replying with pictures,
checking if the command requires admin permissions,
getting the timestamp of the message, making mentions.
Changed the default command prefix for Telegram to /.
Added a clear indication that the original message
is fake for commands received from Web/Android UI.
Bumped text-to-image to 4.1.1 in order to resolve
some dependency errors for canvas with the latest Node.