feat(Telegram), fix(#50): Add "Welcome" feature to Telegram Bot
Added new "setwelcomemessage" to set org-wide welcome
messages to be posted as soon as a new user arrives.
The message is stored in the org settings table.
Created a new Welcomer manager which is capable of posting
the welcome message if it is set.
Extended the Telegram middleware to handle messages
related to new users arrival. Alllow the Welcomer to handle
such cases.
The welcome message can be personalized using the
%username% placeholder inside it.
Restored the alphabetical order of commands in the
command managers.
Fixed some comments.
Fixed a typo bug in the PermissionManagers which prevented
the HelpCommand from checking admin status.